+++ libraries/cloudflare.rb
@@ -7,20 +7,50 @@ end
 
 # adding a few useful methods to the vanilla Cloudflare library
 class CloudflareClient < CloudFlare::Connection
-
     # custom timeout to avoid timing out
     TIMEOUT = 10
 
-    # This method deletes a DNS record by name
+    # This method deletes DNS records by name
     #
     # @param zone [String]
     # @param name [String]
     def rec_delete_by_name zone, name
-        rec_id = get_record(zone, name)['rec_id'] or return false
-        rec_delete(zone, rec_id)
+        get_records(zone, name).keys.each { |rec_id| rec_delete(zone, rec_id) }
     end
 
-    # This method returns true if that zone exists
+    # This method finds a record by its name, zone, content and type,
+    # Returns nil of none was found, the meta hash otherwise
+    #
+    # @param zone [String]
+    # @param name [String]
+    # @param content [String]
+    # @param type [String]
+    def find_single_record zone, name, content, type = 'A'
+        get_records(zone, name).values.each do |rec|
+            if rec['zone_name'] == zone \
+                && rec['display_name'] == name \
+                && rec['content'] == content \
+                && rec['type'] == type
+                return rec
+            end
+        end
+        nil
+    end
+
+    # This method deletes a single record
+    # Returns true iff a record was deleted
+    #
+    # @param zone [String]
+    # @param name [String]
+    # @param content [String]
+    # @param type [String]
+    def rec_delete_single zone, name, content, type = 'A'
+        rec = find_single_record(zone, name, content, type) or return false
+        rec_delete(zone, rec['rec_id'])
+        true
+    end
+
+    # This method returns true iff that zone exists
     #
     # @param zone [String]
     def zone_exists? zone
@@ -35,15 +65,20 @@ class CloudflareClient < CloudFlare::Connection
         zone_load_multi['result'] == 'success' rescue false
     end
 
-    # Gets all the metadata hash for a given record
+    # Gets all the metadata for records that match
+    # the given `zone' and `name'
+    # Returns a hash mapping record's IDs to the metadata hashes
     # 
     # @param zone [String]
     # @param name [String]
-    def get_record zone, name
+    def get_records zone, name
+        result = {}
         get_all_records_for_zone(zone).each do |rec|
-            return rec if rec['display_name'] == name && rec['zone_name'] == zone
-        end rescue NoMethodError
-        nil
+            if rec['display_name'] == name && rec['zone_name'] == zone
+                result[rec['rec_id']] = rec
+            end
+        end
+        result
     end
 
     # The vanilla lib's 'rec_load_all' implementation doesn't account (as of v2.0.1)
@@ -54,6 +89,20 @@ class CloudflareClient < CloudFlare::Connection
     # @param offset [Integer]
     def rec_load_all zone, offset = 0
         send_req({a: :rec_load_all, z: zone, o: offset})
+    end
+
+    # we need to flush the cache when creating, editing or deleting records
+    def rec_new zone, *args
+        flush_cache_for_zone zone
+        super zone, *args
+    end
+    def rec_edit zone, *args
+        flush_cache_for_zone zone
+        super zone, *args
+    end
+    def rec_delete zone, *args
+        flush_cache_for_zone zone
+        super zone, *args
     end
 
     private
@@ -85,4 +134,11 @@ class CloudflareClient < CloudFlare::Connection
         @zone_load_multi_cache ||= super
     end
 
+    # We need to flush the cache when we make changes that make it obsolete
+    #
+    # @param zone [String]
+    def flush_cache_for_zone zone
+        @records_cache ||= {}
+        @records_cache.delete zone
+    end
 end